backend/pyproject.toml
[project]
name = "youtube-sub-manager"
version = "0.1.0"
description = "YouTube subscription video manager backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "httpx>=0.27",
    "google-auth>=2.29",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.130",
    "apscheduler>=3.10",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "defusedxml>=0.7",
    "firebase-admin>=6.5",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx",  # for TestClient
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
